backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
google-cloud-firestore>=2.17.2
google-cloud-storage>=2.18.2
httpx>=0.27.2
python-dotenv>=1.0.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
firebase-admin>=6.5.0
google-genai>=1.0.0
google-cloud-aiplatform>=1.50.0
